LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/transformations/2d.transformation.scale.nonfinite-expected.txt

2d.transformation.scale.nonfinite
scale() with Infinity/NaN is ignored
Actual output:
Expected output:


PASS scale() with Infinity/NaN is ignored 
